requirements.txt
selenium
webdriver-manager
pytest
pyyaml
